% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemLn.R
\docType{methods}
\name{alk}
\alias{alk}
\alias{alk-method}
\alias{alk,FLPar-method}
\title{alk}
\usage{
alk(object, ...)
}
\arguments{
\item{object}{\code{FLPar} with `linf` or  \code{numeric}}

\item{...}{any additional arguments}
}
\description{
alk
}
\details{
alk
}
\examples{
\dontrun{
ak=alk(FLPar(linf=50))
}
}
\seealso{
lenSample
}
\author{
Laurence Kell, Sea++
}
