% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datasets}
\alias{datasets}
\alias{lh}
\alias{prior}
\alias{om}
\alias{eq}
\title{mydas datasets}
\description{
Example datasets for the classes defined in FLCore.
}
\examples{
\dontrun{
data(om)
summary(om)
}
}
\seealso{
\linkS4class{FLStock}, \linkS4class{FLBRP}, \linkS4class{FLPar}
}
\keyword{datasets}
