% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R, R/hcr2o3.R, R/hcrConstantCatch.R,
%   R/hcrParam.R, R/hcrSBTD.R, R/hcrSBTP.R
\docType{methods}
\name{hcr}
\alias{hcr}
\alias{hcr
#'}
\alias{hcr2o3}
\alias{hcrConstantCatch}
\alias{hcrParam}
\alias{hcrSBTD}
\alias{hcrSBTP}
\title{hcr}
\usage{
hcr(object, refs, ...)

hcr2o3(yrs, catch, index, fproxy = 1, bproxy = 1, rule = c("2o3",
  "slope", "loess"))

hcrConstantCatch(yrs, catch, ...)

hcrParam(ftar, btrig, fmin, blim)

hcrSBTD(yrs, control = FLPar(c(k1 = 0.25, k2 = 0.25, gamma = 1)), index,
  catch, ...)

hcrSBTP(yrs, control, ref, target, cpue, catch, wt = 0.5)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{...}{other parameters, i.e.
refs \code{FLPar or FLBRP} object with reference points, can be missing if refpts are part of \code{object}
params \code{FLPar} object with hockey stick HCR parameters, see hcrParam
yr numeric vector with years used to values in HCR
byr numeric vector with years used for bounds
hyr numeric vector with years to use in projection
tac \code{logical} should return value be TAC rather than F?
bndF \code{vector} with bounds (i.e.min and max values) on iter-annual variability on  F
bndTac \code{vector} with bounds (i.e. min and max values) on iter-annual variability on TAC
stkYrs numeric vector with years for calculating stock status, max(as.numeric(dimnames(stock(object))$year)),
refYrs numeric vector with years for calculating reference points, max(as.numeric(dimnames(catch(object))$year)),
hcrYrs numeric vector with years for setting management, max(as.numeric(dimnames(stock(object))$year)),
tacMn \code{logical} should the TACs be the average of the values returned?
maxF  =2 numeric vector specifying maximum relative F}
}
\value{
\code{FLPar} object with value(s) for F or TAC if tac==TRUE
}
\description{
Harvest Control Rule, calculates F, or Total Allowable Catch (TAC) based on a hockey stock harvest control rule.
}
\details{
hcr

hcr

hcr

hcr

hcr
}
\examples{
\dontrun{
bd   =sim()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hcr(bd,yr=i,yr=i+1)$hvt)
}
\dontrun{
data(pl4)
}

\dontrun{
data(pl4)
}

\dontrun{
data(pl4)
}
\dontrun{
data(pl4)
}
\dontrun{
data(pl4)
}
}
\author{
Laurence Kell, Sea++

Laurence Kell, Sea++

Laurence Kell, Sea++

Laurence Kell, Sea++

Laurence Kell, Sea++
}
