% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr2o3.R
\docType{methods}
\name{hcr2o3}
\alias{hcr2o3}
\title{hcr2o3}
\usage{
hcr2o3(yrs, catch, index, fproxy = 1, bproxy = 1, rule = c("2o3",
  "slope", "loess"))
}
\arguments{
\item{yrs}{blah,blah,blah,...}

\item{catch}{blah,blah,blah,...}

\item{index}{blah,blah,blah,...}

\item{fproxy}{1  blah,blah,blah,...}

\item{bproxy}{1 blah,blah,blah,...}

\item{rule}{c("2o3","slope","loess")}
}
\description{
hcr2o3
}
\details{
hcr2o3
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
