% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrParam.R
\docType{methods}
\name{hcrParam}
\alias{hcrParam}
\title{hcrParam}
\usage{
hcrParam(ftar, btrig, fmin, blim)
}
\arguments{
\item{ftar}{blah,blah,blah,...}

\item{btrig}{blah,blah,blah,...}

\item{fmin}{blah,blah,blah,...}

\item{blim}{blah,blah,blah,...}
}
\description{
hcrParam
}
\details{
hcrParam
}
\examples{
\dontrun{
data(pl4)
}
}
\author{
Laurence Kell, Sea++
}
