% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrSBTP.R
\docType{methods}
\name{hcrSBTP}
\alias{hcrSBTP}
\title{hcrSBTP}
\usage{
hcrSBTP(yrs, control, ref, target, cpue, catch, wt = 0.5)
}
\arguments{
\item{yrs}{blah,blah,blah,...}

\item{control}{blah,blah,blah,...}

\item{ref}{blah,blah,blah,...}

\item{target}{blah,blah,blah,...}

\item{cpue}{blah,blah,blah,...}

\item{catch}{blah,blah,blah,...}

\item{wt}{default 0.5  blah,blah,blah,...}
}
\description{
hcrSBTP
}
\details{
hcrSBTP
}
\examples{
\dontrun{
data(pl4)
}
}
\author{
Laurence Kell, Sea++
}
