% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockeyStick.R
\name{hockeyStick}
\alias{hockeyStick}
\alias{hockeyStick-method}
\alias{hockeyStick,FLStock,FLBRP-method}
\alias{hockeyStick,biodyn,FLPar-method}
\title{hockeyStick}
\usage{
hockeyStick(object, refs, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{refs}{\code{FLPar or FLBRP} object with reference points, can be missing if refpts are part of \code{object}
params \code{FLPar} object with hockey stick HCR parameters, see hcrParam
yr numeric vector with years used to values in HCR
byr numeric vector with years used for bounds
hyr numeric vector with years to use in projection
tac \code{logical} should return value be TAC rather than F?
bndF \code{vector} with bounds (i.e.min and max values) on iter-annual variability on  F
bndTac \code{vector} with bounds (i.e. min and max values) on iter-annual variability on TAC
stkYrs numeric vector with years for calculating stock status, max(as.numeric(dimnames(stock(object))$year)),
refYrs numeric vector with years for calculating reference points, max(as.numeric(dimnames(catch(object))$year)),
hcrYrs numeric vector with years for setting management, max(as.numeric(dimnames(stock(object))$year)),
tacMn \code{logical} should the TACs be the average of the values returned?
maxF  =2 numeric vector specifying maximum relative F}

\item{...}{any additional arguments}
}
\value{
\code{FLPar} object with value(s) for F or TAC if tac==TRUE
}
\description{
A Harvest Control Rule (HCR) to calculate F, or Total Allowable Catch (TAC) based 
on a hockey stock harvest control rule.
}
\examples{
\dontrun{
bd   =sim()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hockeyStick(bd,yr=i,yr=i+1)$hvt)
}
}
