% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseXSA.R
\docType{methods}
\name{icesAR}
\alias{icesAR}
\title{icesAR}
\usage{
icesAR(x, ftar = 1, fmin = 0.05, bpa = 0.5, sigma = 0.3)
}
\arguments{
\item{x}{\code{FLBRP} with reference points}

\item{ftar}{default of 1.0}

\item{fmin}{default of0.05}

\item{bpa}{default of0.5}

\item{sigma}{default of 0.3}
}
\description{
# The ICES Advice Rule for stocks where a wide range of stock sizes has been seen 
so the stock recruitment relationship can be estimated

if SSB>= MSY Btrigger
   F=Fmsy
else
   F=FMSY*SSB/Btrigger  

FMSY could be a proxy for FMSY, i.e. F0.1

Where
  Btrigger is BPA,  i.e. Blim×exp(-1.645×sigma)
  Blim is the break point from segmented regression
  
  # http://ices.dk/sites/pub/Publication%20Reports/Advice/2017/2017/12.04.03.01_Reference_points_for_category_1_and_2.pdf
}
\details{
icesAR
}
\examples{
\dontrun{
data(pl4)
}
}
\author{
Laurence Kell, Sea++
}
