% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbspr.R
\docType{methods}
\name{lbspr}
\alias{lbspr}
\alias{lbspr-method}
\alias{lbspr,data.frame,FLPar-method}
\title{lbspr}
\usage{
lbspr(object, params, ...)
}
\arguments{
\item{object}{\code{data.frame} with length frequency sample}

\item{params}{\code{FLPar} with values for linf, a50, ato95 and mk}

\item{...}{any additional arguments}
}
\description{
A wrapper for LBSPR  see https://cran.r-project.org/web/packages/LBSPR/vignettes/LBSPR.html
 This alloe fitting to empirical length data to provide an estimate of the spawning potential 
 ratio (SPR) for use in MSE.
}
\details{
lbspr
}
\examples{
\dontrun{
lfd=lenSample(catch.n(om)[,20:65],ak,nsample=500)
lb=lbspr(lfd,prior) 
}
}
\author{
Laurence Kell, Sea++
}
