% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemLn.R
\docType{methods}
\name{lenSample}
\alias{lenSample}
\title{slenSample}
\usage{
lenSample(object, alk, ...)
}
\arguments{
\item{object}{\code{FLStock}}

\item{alk}{\code{FLPar} with ALK}

\item{...}{additional arguments}
}
\description{
slenSample
}
\details{
lenSample
}
\examples{
\dontrun{
lfd=lenSample(catch.n(ple4)[,ac(2000:2005)],alk,nsample=100)
}
}
\seealso{
setALK
}
\author{
Laurence Kell, Sea++
}
