% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemLn.R, R/mode.R
\docType{functions}
\name{unbin}
\alias{unbin}
\title{mode}
\usage{
unbin(x)

unbin(x)
}
\arguments{
\item{x;}{a vector of with intervals as names}

\item{x;}{a vector of with intervals as names}
}
\value{
a \code{data.frame} with left and right boundaries and mid points.

a \code{data.frame} with left and right boundaries and mid points.
}
\description{
For a vector with labels corresponding to intervals i.e. \code{"(0,10]"}
 returns a data.frame with left and right boundaries and mid point.

For a vector with labels corresponding to intervals i.e. \code{"(0,10]"}
 returns a data.frame with left and right boundaries and mid point.
}
\examples{
x=summary(cut(runif(100),seq(0,1,.1)))
unbin(x)
x=summary(cut(runif(100),seq(0,1,.1)))
unbin(x)
}
