% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlz.R
\docType{methods}
\name{mlzFn}
\alias{mlzFn}
\title{mlz}
\usage{
mlzFn(year, len, lc, linf, k, ncp = 2, ss = 500)
}
\arguments{
\item{year}{blah,blah,blah,...}

\item{len}{blah,blah,blah,...}

\item{lc}{blah,blah,blah,...}

\item{linf}{v blah,blah,blah,...}

\item{k}{blah,blah,blah,...}

\item{ncp}{blah,blah,blah,...}

\item{ss}{blah,blah,blah,...}
}
\description{
mlz
}
\details{
mlz
}
\examples{
\dontrun{
res=mydas:::mlz(mnLen[,ac(40:60)],prior)
}
}
\author{
Laurence Kell, Sea++
}
