% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseXSA.R
\docType{methods}
\name{mseXSA}
\alias{mseXSA}
\alias{icesAR}
\title{mseXSA}
\usage{
icesAR(x, ftar = 1, fmin = 0.05, bpa = 0.5, sigma = 0.3)
}
\arguments{
\item{ftar}{\code{numeric}  default is 1.0}

\item{fmin}{\code{numeric}  default is 0.05 HCR}

\item{bpa}{\code{numeric}  default is 0.5 HCR}

\item{sigma}{\code{numeric}  default is 0.3 HCR}

\item{om}{\code{FLStock}}

\item{eq}{\code{FLBRP}}

\item{mp}{\code{FLStock}}

\item{control}{\code{FLXSA.control}}

\item{rf}{\code{}}

\item{bndTac\code{numeric}}{default is c(0.01,100) Bounds on TAC changes}

\item{start}{\code{numeric}  default is range(om)["maxyear"]-30}

\item{end}{\code{numeric}  default is range(om)["maxyear"]-interval}

\item{interval}{\code{numeric}  default is 3 years over which to run MSE, doesnt work if interval==1, this is a bug}

\item{srDev}{\code{FLQuant} Stochasticity, either by default or supplied as args rlnoise(dim(om)[6],FLQuant(0,dimnames=list(year=start:(end+interval))),0.3),}

\item{uDev}{\code{FLQuant} rlnoise(dim(mp)[6],FLQuant(0,dimnames=dimnames(iter(stock.n(om),1))),0.2),}

\item{maxF}{\code{numeric} 1.0 Capacity, i.e. F in OM can not be greater than this}

\item{whitebox}{\code{boolean} FALSE}
}
\description{
mseXSA
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
