% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSBTDAlb.R
\docType{methods}
\name{mseEMPSBT1Alb}
\alias{mseEMPSBT1Alb}
\title{mseEMPSBT1Alb}
\usage{
mseEMPSBT1Alb(om, eq, control = c(k1 = 2.5, k2 = 2.5, gamma = 1),
  interval = 3, start = range(om)["maxyear"] - 30,
  end = range(om)["maxyear"] - interval, srDev, uDev, selDev,
  maxF = 1.5)
}
\description{
mseEMPSBT1Alb
}
\details{
mseEMPSBT1Alb
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
