% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseMPB.R
\docType{methods}
\name{mseMPB}
\alias{mseMPB}
\alias{mseMPB-method}
\alias{mseMPB,FLStock,FLBRP-method}
\title{mseMPB}
\usage{
mseMPB(om, eq, mp, ftar = 1, btrig = 0.5, fmin = 0.05, blim = 0.3,
  interval = 3, start = range(om)["maxyear"] - 30,
  end = range(om)["maxyear"] - interval, sr_deviances, u_deviances,
  maxF = 1.5)
}
\arguments{
\item{om}{\code{FLStock} object as the operating model}

\item{eq}{blah,blah,blah,...}

\item{mp}{blah,blah,blah,...}

\item{ftar}{1.0  blah,blah,blah,...}

\item{btrig}{0.5  blah,blah,blah,...}

\item{fmin}{0.05  blah,blah,blah,...}

\item{blim}{0.3  blah,blah,blah,...}

\item{interval}{3  blah,blah,blah,...}

\item{start}{\code{numeric}  default is range(om)["maxyear"]-30}

\item{end}{\code{numeric}  default is range(om)["maxyear"]-interval}

\item{sr_deviances}{rlnorm(dim(om)[6],FLQuant(0,dimnames=list(year=start:end)),0.3),}

\item{u_deviances}{=rlnorm(dim(om)[6],FLQuant(0,dimnames=list(year=start:end)),0.2),}

\item{maxF}{1.5}
}
\description{
mseMPB
}
\details{
#http://ices.dk/sites/pub/Publication%20Reports/Advice/2017/2017/12.04.03.01_Reference_points_for_category_1_and_2.pdf
mse
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
