% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSBTD.R
\docType{methods}
\name{mseSBTD}
\alias{mseSBTD}
\title{mseSBTD}
\usage{
mseSBTD(om, eq, control = "missing", srDev, uDev, interval = 1,
  start = range(om)["maxyear"] - 30, end = range(om)["maxyear"] -
  interval, nyrs = 5, cpueFn = ssb, lag = 1, maxF = 10)
}
\description{
mseSBTD
}
\details{
mseSBTD
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
