% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSBTP.R
\docType{methods}
\name{mseSBTP}
\alias{mseSBTP}
\title{mseSBTP}
\usage{
mseSBTP(om, eq, srDev, uDev, interval = 1, start = range(om)["maxyear"]
  - 30, end = range(om)["maxyear"] - interval, control = c(k1 = 0.25,
  k2 = 0.25), yrU = 1:2, yrCatch = seq(interval) - 1,
  refU = "missing", refCatch = "missing", maxF = 1.5)
}
\description{
mseSBTP
}
\details{
mse
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
