% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSRA.R
\docType{methods}
\name{mseSRA}
\alias{mseSRA}
\title{mseSRA}
\usage{
mseSRA(om, eq, mp, ftar = 1, btrig = 0.5, fmin = 0.05, blim = 0.3,
  interval = 3, start = range(om)["maxyear"] - 30,
  end = range(om)["maxyear"] - interval, srDev, maxF = 1.5)
}
\description{
mseSRA
}
\details{
mse
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
