% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseXSA.R
\docType{methods}
\name{mseXSA}
\alias{mseXSA}
\title{mseXSA}
\usage{
mseXSA(om, eq, mp, control, rf = "missing", ftar = 1, fmin = 0.05,
  bpa = 0.5, sigma = 0.3, bndTac = c(0.01, 100), interval = 3,
  start = range(om)["maxyear"] - 30, end = range(om)["maxyear"] -
  interval, sr_deviances = rlnoise(dim(om)[6], FLQuant(0, dimnames =
  list(year = start:(end + interval))), 0.3),
  u_deviances = rlnoise(dim(mp)[6], FLQuant(0, dimnames =
  dimnames(iter(stock.n(om), 1))), 0.2), maxF = 1, whitebox = FALSE)
}
\arguments{
\item{om}{\code{FLStock} object as the operating model}

\item{eq}{\code{FLBRP}   blah,blah,blah,...}

\item{mp}{\code{FLStock}  blah,blah,blah,...}

\item{control}{\code{FLXSA.control}  blah,blah,blah,...}

\item{rf}{\code{}  blah,blah,blah,...}

\item{ftar}{\code{numeric}  default is 1.0}

\item{fmin}{\code{numeric}  default is 0.05 HCR}

\item{bpa}{\code{numeric}  default is 0.5 HCR}

\item{sigma}{\code{numeric}  default is 0.3 HCR}

\item{bndTac}{\code{numeric}  default is c(0.01,100) Bounds on TAC changes}

\item{interval}{\code{numeric}  default is 3 years over which to run MSE, doesnt work if interval==1, this is a bug}

\item{start}{\code{numeric}  default is range(om)["maxyear"]-30}

\item{end}{\code{numeric}  default is range(om)["maxyear"]-interval}

\item{sr_deviances}{\code{FLQuant} Stochasticity, either by default or supplied as args rlnoise(dim(om)[6],FLQuant(0,dimnames=list(year=start:(end+interval))),0.3),}

\item{u_deviances}{\code{FLQuant} rlnoise(dim(mp)[6],FLQuant(0,dimnames=dimnames(iter(stock.n(om),1))),0.2),}

\item{maxF}{\code{numeric} 1.0 Capacity, i.e. F in OM can not be greater than this}

\item{whitebox}{\code{boolean} FALSE}
}
\description{
MSE with XSA as the stock assessment in the management procedure
}
\examples{
\dontrun{
data(pl4)
}
}
\author{
Laurence Kell, Sea++
}
