% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\docType{methods}
\name{survey,FLStock-method}
\alias{survey,FLStock-method}
\alias{oem}
\alias{oem-method}
\alias{oem,FLStock-method}
\title{oem}
\usage{
\S4method{survey}{FLStock}(object, timing = FLQuant(0, dimnames =
  dimnames(m(object))), wt = stock.wt(object), sd = 0, log = FALSE,
  ...)
}
\arguments{
\item{object}{\code{FLStock} for which an index is to be generated}

\item{timing}{blah,blah,blah,...}

\item{...}{any additional arguments}

\item{sel}{\code{FLQuant} vector at age that shapes the catch or biomass.
additive, by default set to \code{TRUE},}

\item{fish.dependent}{\code{logical} that determines whether the index is proportional to the 
stock or catch/effort, by default set to \code{TRUE}}

\item{effort}{\code{character} c("h","f") that determines what proxy to use for effort,
i.e. harvest rate or instanteous fishing mortality}

\item{mass}{\code{logical} default is TRUE, whether index is in mass or numbers}
}
\value{
\code{FLQuant} with time series

\code{FLQuant} with time series
}
\description{
Creates an \code{FLQuant} to represent an index of relative abundance
}
\examples{
\dontrun{
 data(ple4) 
 cpue <- oem(ple4) 
 }

\dontrun{
 data(ple4) 
 cpue=oem(ple4) 
 }
}
