% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\docType{methods}
\name{popdyn}
\alias{popdyn}
\title{popdyn}
\usage{
popdyn(object, eq = lhEql(lhPar(object)))
}
\arguments{
\item{object}{blah,blah,blah,...}

\item{eq}{blah,blah,blah,...}
}
\description{
Calculates popdyn, i.e. r and reference points, based on life history parameters
}
\details{
popdyn
}
\examples{
\dontrun{
par=FLPar(linf=100)
popdyn(par)
}
}
\author{
Laurence Kell, Sea++
}
