% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\docType{methods}
\name{priors}
\alias{priors}
\title{priors}
\usage{
priors(object, eq = lhEql(lhPar(object)))
}
\arguments{
\item{x}{}
}
\description{
calculates priors
}
\details{
priors
}
\examples{
\dontrun{
par=FLPar(linf=100)
priors(par)
}
}
\author{
Laurence Kell, Sea++
}
