% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTime.R
\docType{methods}
\name{gt}
\alias{gt}
\alias{gt-method}
\alias{gt,FLBRP-method}
\title{gt}
\usage{
gt(object, ...)
}
\arguments{
\item{object}{\code{FLBRP}}

\item{...}{other arguments}
}
\description{
Calculates the generation time
}
\details{
gt
}
\examples{
\dontrun{
data(pl4)
alk=gt(FLBRP(ple4))
}
}
\author{
Laurence Kell, Sea++
}
