% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrConstantCatch.R
\docType{methods}
\name{hcrConstantCatch}
\alias{hcrConstantCatch}
\title{hcrConstantCatch}
\usage{
hcrConstantCatch(yrs, catch, ...)
}
\arguments{
\item{yrs}{\code{numeric} target years to take catch over}

\item{catch}{an \code{FLQuant} with catch to average over by iter}

\item{...}{any additional arguments}
}
\description{
Sets a constant catch
}
\details{
hcrConstantCatch
}
\examples{
\dontrun{
data(pl4)
hcrConstantCatch(2010:2012,catch(ple4))
}

}
\author{
Laurence Kell, Sea++
}
