% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcrSBTD.R
\docType{methods}
\name{hcrSBTD}
\alias{hcrSBTD}
\title{hcrSBTD}
\usage{
hcrSBTD(yrs, control = FLPar(c(k1 = 0.25, k2 = 0.25, gamma = 1)), index,
  catch, ...)
}
\arguments{
\item{yrs}{blah,blah,blah,...}

\item{control}{blah,blah,blah,...FLPar(c(k1=0.25,k2=0.25,gamma=1))}

\item{index}{blah,blah,blah,...}

\item{catch}{blah,blah,blah,...}

\item{...}{any additional arguments}
}
\description{
An emprirical Harvest Control Rule that sets a quota based on a trend in an index
}
\details{
hcrSBTD
}
\examples{
\dontrun{
data(pl4)
}
}
\author{
Laurence Kell, Sea++
}
