% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbspr.R
\docType{methods}
\name{lbsprFn}
\alias{lbsprFn}
\title{lbspr}
\usage{
lbsprFn(len, params, species = "", units = "cm")
}
\arguments{
\item{x}{}
}
\description{
lbspr
}
\details{
lbspr
}
\examples{
\dontrun{

}
}
\author{
Laurence Kell, Sea++
}
