% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemLn.R
\docType{methods}
\name{lenSample}
\alias{lenSample}
\alias{lenSample-method}
\alias{lenSample,FLQuant,FLPar,missing-method}
\alias{lenSample,FLQuant,FLPar,numeric-method}
\title{lenSample}
\usage{
lenSample(object, alk, nsample, ...)
}
\arguments{
\item{object}{\code{FLStock} blah,blah,blah,...}

\item{alk}{\code{FLPar} with ALK}

\item{nsample}{\code{numeric} sample size, default is 500}

\item{...}{any additional arguments}
}
\description{
lenSample
}
\details{
lenSample
}
\examples{
\dontrun{
lfd=lenSample(catch.n(ple4)[,ac(2000:2005)],alk,nsample=100)
}
}
\seealso{
setALK
}
\author{
Laurence Kell, Sea++
}
