% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlz.R
\docType{methods}
\name{mlzFn}
\alias{mlzFn}
\title{mlz}
\usage{
mlzFn(year, len, lc, linf, k, ncp = 2, ss = 500)
}
\arguments{
\item{x}{}
}
\description{
mlz
}
\details{
mlz
}
\examples{
\dontrun{

}
}
\author{
Laurence Kell, Sea++
}
