% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSBTDAlb.R
\docType{methods}
\name{mseEMPSBT1Alb}
\alias{mseEMPSBT1Alb}
\title{mseEMPSBT1Alb}
\usage{
mseEMPSBT1Alb(om, eq, control = c(k1 = 2.5, k2 = 2.5, gamma = 1),
  interval = 3, start = range(om)["maxyear"] - 30,
  end = range(om)["maxyear"] - interval, sr_deviances, u_deviances,
  sel_deviances, maxF = 1.5)
}
\arguments{
\item{om}{\code{FLStock} object as the operating model}

\item{eq}{blah,blah,blah,...}

\item{control}{blah,blah,blah,...}

\item{interval}{blah,blah,blah,...}

\item{start}{\code{numeric}  default is range(om)["maxyear"]-30}

\item{end}{\code{numeric}  default is range(om)["maxyear"]-interval}

\item{sr_deviances}{blah,blah,blah,...}

\item{u_deviances}{blah,blah,blah,...}

\item{sel_deviances}{blah,blah,blah,...}

\item{maxF}{blah,blah,blah,...}
}
\description{
mseEMPSBT1Alb
}
\details{
mseEMPSBT1Alb
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
