% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSBTPAlb.R
\docType{methods}
\name{mseEMPSBT2Alb}
\alias{mseEMPSBT2Alb}
\title{mseEMPSBT2Alb}
\usage{
mseEMPSBT2Alb(om, eq, control = c(k1 = 0.25, k2 = 0.25),
  refYr = "missing", interval = 3, start = range(om)["maxyear"] - 30,
  end = range(om)["maxyear"] - interval, sr_deviances, u_deviances,
  sel_deviances, maxF = 1.5)
}
\arguments{
\item{om}{\code{FLStock} object as the operating model}

\item{eq}{blah,blah,blah,...}

\item{control}{blah,blah,blah,...}

\item{refYr}{blah,blah,blah,...}

\item{interval}{blah,blah,blah,...}

\item{start}{\code{numeric}  default is range(om)["maxyear"]-30}

\item{end}{\code{numeric}  default is range(om)["maxyear"]-interval}

\item{sr_deviances}{blah,blah,blah,...}

\item{u_deviances}{blah,blah,blah,...}

\item{sel_deviances}{blah,blah,blah,...}

\item{maxF}{blah,blah,blah,...}
}
\description{
A function to run a full feedback MSE, using an emprirical HCR
}
\details{
mse
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
