% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseSBTPAlb.R
\docType{methods}
\name{mseEMPSBT2Alb}
\alias{mseEMPSBT2Alb}
\title{mseEMPSBT2Alb}
\usage{
mseEMPSBT2Alb(om, eq, control = c(k1 = 0.25, k2 = 0.25),
  refYr = "missing", interval = 3, start = range(om)["maxyear"] - 30,
  end = range(om)["maxyear"] - interval, srDev, uDev, selDev,
  maxF = 1.5)
}
\description{
mseEMPSBT2Alb
}
\details{
mse
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
