% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseMPB.R
\docType{methods}
\name{mseMPB}
\alias{mseMPB}
\alias{mpTrace}
\alias{mseMPB-method}
\alias{mseMPB,FLStock,FLBRP-method}
\title{mseMPB}
\usage{
mpTrace(mp, par, tac)
}
\arguments{
\item{blim=0.3, }{}

\item{#years}{over which to run MSE}

\item{end=range(om)["maxyear"]-interval, }{#Stochasticity}

\item{srDev}{#=rlnorm(dim(om)[6],FLQuant(0,dimnames=list(year=start:end)),0.3),}

\item{uDev}{#=rlnorm(dim(om)[6],FLQuant(0,dimnames=list(year=start:end)),0.2),
#Capacity, i.e. F in OM can not be greater than this}
}
\description{
mseMPB
}
\details{
#http://ices.dk/sites/pub/Publication%20Reports/Advice/2017/2017/12.04.03.01_Reference_points_for_category_1_and_2.pdf
mse
}
\examples{
\dontrun{
data(pl4)
}

}
\author{
Laurence Kell, Sea++
}
