% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omOut.R
\docType{methods}
\name{omSmry}
\alias{omSmry}
\title{omSmry}
\usage{
omSmry(x, ...)
}
\arguments{
\item{x}{\code{FLStock}}

\item{...}{any other arguments}

\item{y}{\code{FLBRP} or \code{FLPar} with reference points, optional}

\item{z}{\code{FLPar} with `a` and `b` of length weight relationship \code{FLPar(a-0.001,b=3)} by default}
}
\description{
create time series of summary statistics from \code{FLStock} relative to reference points
}
\details{
omSmry
}
\examples{
\dontrun{
res=omSmry(om,eq)
}
}
\seealso{
smryStat
}
\author{
Laurence Kell, Sea++
}
