% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\docType{methods}
\name{priors}
\alias{priors}
\title{priors}
\usage{
priors(object, eq = lhEql(lhPar(object)))
}
\arguments{
\item{object}{blah,blah,blah,...}

\item{eq}{blah,blah,blah,...}
}
\description{
Calculates priors, i.e. r and reference points, based on life history parameters
}
\details{
priors
}
\examples{
\dontrun{
par=FLPar(linf=100)
priors(par)
}
}
\author{
Laurence Kell, Sea++
}
